requirements.txt
blessed==1.17.9
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
idna==2.10
importlib-metadata==1.7.0
keyring==21.3.1
lxml==4.5.2
measurement==3.2.0
mpmath==1.1.0
myfitnesspal==1.16.1
pprintpp==0.4.0
Pygments==2.6.1
python-dateutil==2.8.1
requests==2.24.0
rich==2.3.1
six==1.15.0
sympy==1.6.2
typing-extensions==3.7.4.2
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0